Refactor livekit realtime as stanadalone typescript
